Split core utils, add tests, rename weekday -> dayOfWeek
